+++ src/components/App.js
@@ -6,7 +6,7 @@ import RegisterForm from "./RegisterForm";
 import Navigation from "./Navigation";
 import Home from "./Home";
 import Footer from "./Footer";
-import Cart from "./Cart";
+import Cart from "./Order/Cart";
 // getAPIHealth is defined in our axios-services directory index.js
 // you can think of that directory as a collection of api adapters
 // where each adapter fetches specific info from our express server's /api route
