+++ src/components/Order/SingleCartProduct.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+const SingleCartProduct = () => {
+  return (
+    <>
+      <div className="title"> product title </div>
+      <div className="title"> price </div>
+      <div> quantity button </div>
+      <div> delete button </div>
+    </>
+  );
+};
+
+export default SingleCartProduct;
