+++ src/components/index.js
@@ -1,6 +1,6 @@
 export { default as App } from './App';
 export { default as Reviews } from './Reviews'
-export { default as Cart } from "./Cart";
+export { default as Cart } from "./Order/Cart";
 export { default as MyAccount } from "./MyAccount/MyAccount";
 export { default as MyOrders } from "./MyAccount/MyOrders";
 export { default as SingleOrder } from "./MyAccount/SingleOrder";
