+++ src/style/Cart.css
@@ -1,4 +1,11 @@
+h2 {
+  margin-top: 15vh;
+}
+
 .checkout {
-    margin-top: 30vh;
-  }
-  
+  margin-top: 15vh;
+}
+
+.title {
+  margin: 5vh;
+}
